common: scpi: Shrink some structure members
Since the maximum SCPI message size is 256 bytes, the payload size will
always fit in a uint8_t. This saves 4 bytes per implemented SCPI
command.

+++ common/scpi_cmds.c
@@ -33,9 +33,9 @@ struct scpi_cmd {
 	/** Fixed reply payload. */
 	uint32_t *tx_payload;
 	/** Expected size of received payload. */
-	uint16_t  rx_size;
+	uint8_t   rx_size;
 	/** Size of fixed reply payload, if present. */
-	uint16_t  tx_size;
+	uint8_t   tx_size;
 	/** Any combination of flags from above, if applicable. */
 	uint8_t   flags;
 };
